requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
pylibdmtx>=0.1.10
Pillow>=10.0.0 